+++ track_converter/tests/test_preprocess_ctc.py
@@ -1,26 +1,11 @@
 import re
-from contextlib import nullcontext as does_not_raise
-from pathlib import Path
 
 import numpy as np
 import pandas as pd
 import pytest
 
-from track_converter.src.preprocess_ctc import discard_related_cells, preprocess_ctc_file
-
-
-@pytest.fixture
-def test_data_dir():
-    return Path(__file__).parent.resolve() / "data" / "CTC"
-
-
-def read_tracks(tracks_path: Path):
-    tracks = pd.read_table(tracks_path, sep=r"\s+", header=None)
-    if len(tracks.columns) == 4:
-        tracks.columns = ["L", "B", "E", "P"]
-    elif len(tracks.columns) == 5:
-        tracks.columns = ["L", "B", "E", "P", "R"]
-    return tracks
+from track_converter.src.preprocess_ctc import preprocess_ctc_file
+from track_converter.tests.utils import read_tracks
 
 
 @pytest.mark.parametrize(
@@ -62,13 +47,13 @@ def read_tracks(tracks_path: Path):
         ),
     ],
 )
-def test_validate_tracks(tracks_file, expected_warning, expected_remaining_labels, tmp_path, test_data_dir, caplog):
+def test_validate_tracks(tracks_file, expected_warning, expected_remaining_labels, tmp_path, ctc_test_data_dir, caplog):
     """
     Test validation of tracks files with various issues.
 
     For each, check the correct cells are removed and the correct warning is given.
     """
-    tracks_in_path = test_data_dir / tracks_file
+    tracks_in_path = ctc_test_data_dir / tracks_file
     tracks_out_path = tmp_path / "tracks_out.txt"
     preprocess_ctc_file(tracks_in_path, tracks_out_path)
 
@@ -86,50 +71,12 @@ def test_validate_tracks(tracks_file, expected_warning, expected_remaining_label
             assert remaining_labels.isin(expected_remaining_labels).all()
 
 
-def test_valid_file_passes(tmp_path, test_data_dir):
+def test_valid_file_passes(tmp_path, ctc_test_data_dir):
     """Run a larger valid example file through the validation, checking no exceptions are raised."""
-    tracks_in_path = test_data_dir / "tracks_valid_large.txt"
+    tracks_in_path = ctc_test_data_dir / "tracks_valid_large.txt"
     tracks_out_path = tmp_path / "tracks_out.txt"
 
     preprocess_ctc_file(tracks_in_path, tracks_out_path)
-
-
-@pytest.mark.parametrize(
-    "tracks_file,labels_to_discard,expected_remaining_labels,expected_exception",
-    [
-        # tracks where all cells are connected in one 'tree' with structure:
-        #    1
-        #   / \
-        #  2   3
-        #     / \
-        #    4   5
-        pytest.param(
-            "tracks_one_tree.txt",
-            (5, 2),
-            (),
-            pytest.raises(ValueError, match=r"No tracks remaining after discarding related cells of \(5, 2\)"),
-            id="one tree",
-        ),
-        # tracks where cells are connected into two 'trees' with structure:
-        #    1       4
-        #   / \     / \
-        #  2   3   5   6
-        pytest.param("tracks_two_trees.txt", (2,), (4, 5, 6), does_not_raise(), id="two trees"),
-    ],
-)
-def test_discard_related_cells(
-    tracks_file, labels_to_discard, expected_remaining_labels, expected_exception, test_data_dir
-):
-    """Test discarding all related cells in different cell lineages."""
-    tracks = read_tracks(test_data_dir / tracks_file)
-
-    with expected_exception:
-        tracks = discard_related_cells(tracks, labels_to_discard)
-        remaining_labels = tracks["L"]
-
-        assert len(remaining_labels) == len(expected_remaining_labels)
-        if len(expected_remaining_labels) > 0:
-            assert remaining_labels.isin(expected_remaining_labels).all()
 
 
 @pytest.mark.parametrize(
@@ -173,8 +120,8 @@ def test_discard_related_cells(
         ),
     ],
 )
-def test_fix_late_daughters(tracks_file, expected_warning, expected_output, tmp_path, test_data_dir, caplog):
-    tracks_in_path = test_data_dir / tracks_file
+def test_fix_late_daughters(tracks_file, expected_warning, expected_output, tmp_path, ctc_test_data_dir, caplog):
+    tracks_in_path = ctc_test_data_dir / tracks_file
     tracks_out_path = tmp_path / "tracks_out.txt"
     preprocess_ctc_file(tracks_in_path, tracks_out_path, fix_late_daughters=True)
 
@@ -230,8 +177,8 @@ def test_fix_late_daughters(tracks_file, expected_warning, expected_output, tmp_
         ),
     ],
 )
-def test_fix_missing_daughters(tracks_file, expected_warning, expected_output, tmp_path, test_data_dir, caplog):
-    tracks_in_path = test_data_dir / tracks_file
+def test_fix_missing_daughters(tracks_file, expected_warning, expected_output, tmp_path, ctc_test_data_dir, caplog):
+    tracks_in_path = ctc_test_data_dir / tracks_file
     tracks_out_path = tmp_path / "tracks_out.txt"
     preprocess_ctc_file(tracks_in_path, tracks_out_path, fix_missing_daughters=True)
